requirements.txt
environs~=9.5.0

fastapi~=0.103.1
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
pydantic[email]~=2.3.0

uvicorn~=0.23.2

alembic~=1.12.0
sqlalchemy[postgresql_asyncpg]~=2.0.21
marshmallow-sqlalchemy~=0.29.0
